Updated readme.md to learn about github commits.
This update was created on a new branch called readme-edits. First the branch was created and initially it looked exactly like the main branch. But now as I have updated the readme.md file, this branch will differ from main and the change will only be committed to this particular branch.